WIP add tabs;navigate back/next;refactor search
