+++ public_src/components/route-modal/route-modal.component.less
@@ -2,7 +2,7 @@
   position: absolute;
   //top: 50px;
   //bottom: 0;
-  width: 60%;
+  width: 100%;
   height: 95%;
 }
